for #2084, move mysql ddl extractor package
